:bug: Broken cookie parsing
CookieSpec handles Set-Cookie headers sent by servers, not Cookie
headers sent by clients. CookieHandler gives us a Cookie header, so
we need an appropriate parser for it.

Tomcat has a solution for this, but it's pretty tangly:
[Rfc6265CookieProcessor handles Cookie headers](https://github.com/apache/tomcat/blob/5d7a579366d48668dd81934a74478336dab06ac7/java/org/apache/tomcat/util/http/Rfc6265CookieProcessor.java)
[Cookie does the low-level processing](https://github.com/apache/tomcat/blob/5d7a579366d48668dd81934a74478336dab06ac7/java/org/apache/tomcat/util/http/parser/Cookie.java)

I'm going to stick with a hand-rolled parser for this for now as it
works.