Fix #138: Update Jsoup to 1.11.3 (#139)
* Fix #138: Update Jsoup to 1.11.3

On OSX, users sometimes get an empty role list while Linux users do not
esperience problems. This holds even true when the same Docker image is
used.

We tracked this down to the document in `AwsSamlSigninParser` (the HTML document
that represents the HTML page AWS presents for selecting roles after login to
Okta) being broken judging by string representation of the `Document` object.
However, we noticed when converting the HTML response to a string before
passing it to `Jsoup.parse`, the error disappeared:

```java
    // in AwsSamlRoleUtils.getSigninPageDocument
    Scanner s = new Scanner(samlSigninResponse.getEntity().getContent()).useDelimiter("\\A");
    String result = s.hasNext() ? s.next() : "";
    System.out.println(result);
    return Jsoup.parse(
	    result,
	    "https://signin.aws.amazon.com/saml"
    );
```

This lead us to the conclusion that this might be a race condition that
only happens when Jsoup is passed an `InputStream` which could explain why it
only happens on OSX and not Linux: Such things can propagate down right to the
operating system through boundaries of VM/Docker.

Judging from Jsoup's changelog from 1.11.2 to 1.11.3, there is a
particular [change](https://github.com/jhy/jsoup/issues/995) that sounds
like this could be the solution. Anyways, when updating to the latest
version, the error disappears completely on our OSX machines.

* Pom: Do not include full AWS SDK

File shrunk by 80MB (99 -> 19)